requirements.txt
opencv-python
mediapipe
win10toast-click
python-dotenv
google
PIL
